Replace string comparators with LIKE, add support for regexp (with RLIKE)
Removes BEGINSWITH/ENDSWITH/CONTAINS/IS in favour of LIKE:

  BEGINSWITH <value> -> LIKE <value>%
  ENDWITH <value> -> LIKE %<value>
  CONTAINS <value> -> LIKE %<value>%
  IS <value> -> LIKE <value>

Introduces RLIKE for regular expression pattern matching. Uses
regexp.MustCompile (see compare.Alpha) which panics on invalid regex. I'm
thinking this should be changed to use regexp.Compile (which returns the
error) and then on error, we fatal with a "nicer" message (although this might
be more work than it sounds, since we'll have to rewrite all comparison
functions to return an error).