Add subquery tokenizer
The currently solution requires us to track both the tokenizer's "previous"
Token as well as each Token's preceding Token.

When we reach an opening paren, we check if it's preceding token is IN, and if
so, the next block is read as a Subquery.
This is just a string representing another input, so it'll be tokenized,
parsed, and evalulated as an individual query.

If the subquery includes parens, this method of reading breaks, since we stop
reading as soon as we reach a closing paren. So, we'll have to count opening
parens and only exit when the count is down to 0.