Output attributes based on order in SELECT clause.
Uses a slice to track attributes instead of a map. Probably takes a
minor toll on performance (unnoticeable as far as I can tell), but it's
the only way to track/maintain order (although we can use both a
slice and a map, but that'll get kind of messy).

Introduces a new step prior to outputting the result to find the
longest name value (for consistent output padding).

Closes #34.